backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
coverage==7.3.2
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.3
django-ratelimit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
factory-boy==3.3.0
Faker==37.4.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Pillow==10.1.0
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
